refactor the fetching
